[perf] reduce sampler overhead a little bit
git-svn-id: https://spf4j.googlecode.com/svn/trunk@269 d408096a-aa36-d13b-a661-48d75a76e993

+++ spf4j-core/src/main/java/org/spf4j/stackmonitor/SamplerMBean.java
@@ -42,11 +42,11 @@ public interface SamplerMBean extends ReportGenerator {
 
     long getSampleTimeMillis();
     
-    void setSampleTimeMillis(long sampleTimeMillis);
+    void setSampleTimeMillis(int sampleTimeMillis);
 
     long getDumpTimeMillis();
     
-    void setDumpTimeMillis(long dumpTimeMillis);
+    void setDumpTimeMillis(int dumpTimeMillis);
     
     boolean isStopped();
